+++ component/language/parser/ast.go
@@ -106,6 +106,8 @@ type (
 	FunctionArg           = ast.FunctionArg
 	AutomationDef         = ast.AutomationDef
 	TriggerDef            = ast.TriggerDef
+	LoopDef               = ast.LoopDef
+	ModeDef               = ast.ModeDef
 	ArgsSchema            = ast.ArgsSchema
 	ArgsField             = ast.ArgsField
 	StepDef               = ast.StepDef
@@ -163,7 +165,9 @@ var (
 	AttrFilter               = ast.AttrFilter
 	AttrHandler              = ast.AttrHandler
 	AttrIdempotent           = ast.AttrIdempotent
+	AttrLoop                 = ast.AttrLoop
 	AttrMergeFields          = ast.AttrMergeFields
+	AttrMode                 = ast.AttrMode
 	AttrAppendFields         = ast.AttrAppendFields
 	AttrAddToSet             = ast.AttrAddToSet
 	AttrRemoveFromSet        = ast.AttrRemoveFromSet